in hash calc, delete old cache files that will not be used earlier (#33432)
* in hash calc, delete old cache files that will not be used earlier

* only delete if supposed to

* fmt

+++ accounts-db/src/cache_hash_data.rs
@@ -198,9 +198,7 @@ pub(crate) struct CacheHashData {
 
 impl Drop for CacheHashData {
     fn drop(&mut self) {
-        if self.should_delete_old_cache_files_on_drop {
-            self.delete_old_cache_files();
-        }
+        self.delete_old_cache_files();
         self.stats.report();
     }
 }
@@ -224,18 +222,24 @@ impl CacheHashData {
         result.get_cache_files();
         result
     }
-    fn delete_old_cache_files(&self) {
-        let old_cache_files = std::mem::take(&mut *self.pre_existing_cache_files.lock().unwrap());
-        if !old_cache_files.is_empty() {
-            self.stats
-                .unused_cache_files
-                .fetch_add(old_cache_files.len(), Ordering::Relaxed);
-            for file_name in old_cache_files.iter() {
-                let result = self.cache_dir.join(file_name);
-                let _ = fs::remove_file(result);
+
+    /// delete all pre-existing files that will not be used
+    pub(crate) fn delete_old_cache_files(&self) {
+        if self.should_delete_old_cache_files_on_drop {
+            let old_cache_files =
+                std::mem::take(&mut *self.pre_existing_cache_files.lock().unwrap());
+            if !old_cache_files.is_empty() {
+                self.stats
+                    .unused_cache_files
+                    .fetch_add(old_cache_files.len(), Ordering::Relaxed);
+                for file_name in old_cache_files.iter() {
+                    let result = self.cache_dir.join(file_name);
+                    let _ = fs::remove_file(result);
+                }
             }
         }
     }
+
     fn get_cache_files(&self) {
         if self.cache_dir.is_dir() {
             let dir = fs::read_dir(&self.cache_dir);
